# Improve eager-loading on admin products page