--- a/app/controllers/api/v0/products_controller.rb
+++ b/app/controllers/api/v0/products_controller.rb
@@ -116,7 +116,7 @@ module Api
 
       def product_query_includes
         [
-          master: [:images],
+          master: { images: { attachment_attachment: :blob } },
           variants: [:default_price, :stock_locations, :stock_items, :variant_overrides,
                      { option_values: :option_type }]
         ]