--- a/app/serializers/api/admin/product_serializer.rb
+++ b/app/serializers/api/admin/product_serializer.rb
@@ -5,12 +5,26 @@ module Api
     class ProductSerializer < ActiveModel::Serializer
       attributes :id, :name, :sku, :variant_unit, :variant_unit_scale, :variant_unit_name,
                  :inherits_properties, :on_hand, :price, :available_on, :permalink_live,
-                 :tax_category_id, :import_date, :image_url, :thumb_url
+                 :tax_category_id, :import_date, :image_url, :thumb_url, :variants, :master
 
       has_one :supplier, key: :producer_id, embed: :id
       has_one :primary_taxon, key: :category_id, embed: :id
-      has_many :variants, key: :variants, serializer: Api::Admin::VariantSerializer
-      has_one :master, serializer: Api::Admin::VariantSerializer
+
+      def variants
+        ActiveModel::ArraySerializer.new(
+          object.variants,
+          each_serializer: Api::Admin::VariantSerializer,
+          image: thumb_url,
+          stock_location: Spree::StockLocation.first
+        )
+      end
+
+      def master
+        Api::Admin::VariantSerializer.new(
+          object.master,
+          image: thumb_url
+        )
+      end
 
       def image_url
         object.images.first&.url(:product) || "/noimage/product.png"