#### What? Why?

For some reason eager-loaded data (using `includes`) isn't passed down to nested serializers when using associations. Defining the association as a method and passing in some optional paramaters resolves the issue.

This reduces queries run by the controller action by around 30%, and has a decent effect on the response times.


#### What should we test?

Admin products page, loading the products should be significantly faster.

#### Release notes

<!-- Please select one for your PR and delete the other. -->

Changelog Category: Technical changes

<!-- Choose a pull request title above which explains your change to a
     a user of the Open Food Network app. -->

The title of the pull request will be included in the release notes.