@drummer83 2023-05-17
Hi @Matt-Yorkley,
I love the fact that you are working on performance issues. Thank you so much for that!

I compared the loading times of admin/products before and after using the performance tool of Chromium.
- Scripting is quite a bit faster now: 4.2 s --> 3.5 s
- System and Idle have increased: 325 ms --> 577 ms and 629 ms --> 1230 ms
- Total time is more or less identical.

Before:
![image](https://github.com/openfoodfoundation/openfoodnetwork/assets/1790176/15e2de91-0a95-4f3c-b8ff-ce0eb33dfafa)

After: 
![image](https://github.com/openfoodfoundation/openfoodnetwork/assets/1790176/f6496d44-ea1d-49c4-bf85-b57ff632c22b)

The page is loading properly, so I think we are good here. 
I will leave it to you to merge if you are happy with this. Thanks!