@Matt-Yorkley 2023-05-17
Ah I might have been looking at this on a branch combined with #10812 previously. Here's some new stats, the effect is a bit less (~30% reduction) but still a good improvement. It's on the server side, so the profiling in chrome isn't a useful tool to assess it.

**Before:**
![Screenshot from 2023-05-17 18-42-38](https://github.com/openfoodfoundation/openfoodnetwork/assets/9029026/de9c69bb-5925-4efb-8f8c-7e8bfa08c3ca)

**After:**
![Screenshot from 2023-05-17 18-41-45](https://github.com/openfoodfoundation/openfoodnetwork/assets/9029026/43b1e494-76cd-4d70-bdea-725e90a2eb6c)

The number of queries here is still crazy but it's a step in the right direction :+1: